change app struct, handle catch-all package, etc...
