+++ README.md
@@ -1,8 +1,4 @@
-В соответствии с итоговым заданием реализован проект небольшого сервиса по управлению задачами.
-Проект писался в течение двух недель с перерывами и за промежутки отдыха я несколько раз пересматривал свой подход.
-Поэтому, как мне кажется, единая стилистика не выдержана.
-Я всё еще вижу много вещей, которые можно переписать и улучшить, но в данный момент я больше не готов самостоятельно
-заниматься перфекционизмом :).
+В соответствии с итоговым заданием реализован проект небольшого сервиса планирования задач.
 
 Тесты запускаются командой: 
 go test -count=1 ./tests
@@ -32,3 +28,5 @@ var Token = `какой-то токен`
 
 docker run -d -p 7540:7540 -v /path/to/db:/app/data -e TODO_PORT=:7540 -e TODO_DBFILE=/app/data/scheduler.db
 -e TODO_PASSWORD="12345" -e TODO_SECRET="my_secret_key" my-todo-app
+
+В версии после первой проверки полностью переработана структура проекта.
