+++ internal/constants/constants.go
@@ -1,0 +1,6 @@
+package constants
+
+const (
+	DateFormat = "20060102"
+	TasksLimit = 10
+)
