--- pkg/constants/constants.go
@@ -1,3 +1,0 @@
-package constants
-
-const DateFormat = "20060102"
