+++ tests/settings.go
@@ -4,4 +4,4 @@ var Port = 7540
 var DBFile = "../scheduler.db"
 var FullNextDate = false
 var Search = true
-var Token = ``
+var Token = `eyJhbGciOiJIUzI1NiIsInR5cCI6IkpXVCJ9.eyJleHAiOjE3NDI2ODA1NzIsInB3ZCI6IjU5OTQ0NzFhYmIwMTExMmFmY2MxODE1OWY2Y2M3NGI0ZjUxMWI5OTgwNmRhNTliM2NhZjVhOWMxNzNjYWNmYzUifQ.AjRZqWQIoT62bDNzCGca__lx4T8n7hwr73_SeO95ay8`
